Move impl to separate files
